backend/requirements.txt
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
django-filter==25.1
djangorestframework-simplejwt==5.4.0
stripe==12.2.0
